@@ -3,6 +3,7 @@
   "scripts": {
     "lint": "eslint --ext .js,.ts .",
     "build": "tsc",
+    "build:watch": "tsc --watch",
     "serve": "npm run build && firebase emulators:start --only functions",
     "shell": "npm run build && firebase functions:shell",
     "start": "npm run shell",
@@ -12,21 +13,21 @@
     "test:converter": "firebase emulators:exec --only firestore,database --import=./../server_data \"jest converter._jest.spec.js\"",
     "emulateSnapshot": "firebase emulators:start --only firestore,database --import=./../server_data "
   },
-  "main": "index.js",
+  "main": "lib/index.js",
   "dependencies": {
     "@google-cloud/pubsub": "^2.18.5",
-    "firebase-admin": "^9.11.0",
+    "firebase-admin": "^10.0.2",
     "firebase-functions": "^3.17.1"
   },
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^3.9.1",
-    "@typescript-eslint/parser": "^3.8.0",
+    "@typescript-eslint/eslint-plugin": "^5.10.2",
+    "@typescript-eslint/parser": "^5.10.2",
+    "eslint": "^8.8.0",
     "eslint-config-google": "^0.14.0",
     "eslint-plugin-import": "^2.22.0",
-    "typescript": "^3.8.0",
-    "eslint": "^7.32.0",
-    "eslint-plugin-promise": "^4.3.1",
-    "firebase-functions-test": "^0.3.1"
+    "eslint-plugin-promise": "^6.0.0",
+    "firebase-functions-test": "^0.3.1",
+    "typescript": "^4.5.5"
   },
   "private": true,
   "engines": {
