@@ -1,5 +1,4 @@
-const functions = require("firebase-functions");
-const admin = require("firebase-admin");
+import * as functions from "firebase-functions";
 
 const environment = require("../environments/environment.js");
 
@@ -7,8 +6,7 @@ const onChange = require("./members/onChange.js");
 const copy = require("./members/copy.js");
 
 const context = {
-  admin,
-  environment
+  environment,
 };
 
 exports.onChange = functions.firestore
