@@ -1,5 +1,4 @@
-const functions = require("firebase-functions");
-const admin = require("firebase-admin");
+import * as functions from "firebase-functions";
 
 const environment = require("../environments/environment.js");
 
@@ -8,8 +7,7 @@ const updateStats = require("./trees/updateStats.js");
 const onChange = require("./trees/onChange.js");
 
 const context = {
-  admin,
-  environment
+  environment,
 };
 
 exports.rebuildTree = functions.firestore
