@@ -33,50 +33,58 @@
 </template>
 
 <script>
-import { mapGetters, mapState } from "vuex";
-import { useQuasar } from "quasar";
-import { getFirestore, setDoc, doc } from "@firebase/firestore";
+import { mapGetters, mapState } from 'vuex';
+import { useQuasar } from 'quasar';
+import { getFirestore, setDoc, doc, arrayUnion } from '@firebase/firestore';
 
 export default {
-  name: "AddSubTree",
-  props: ["parent", "treeId"],
+  name: 'AddSubTree',
+  props: ['parent', 'treeId'],
   setup() {
     const q = useQuasar();
     return { q };
   },
   data() {
     return {
-      subTree: "",
+      subTree: '',
       shadow: false,
     };
   },
   methods: {
     addTree() {
-      // console.log(
-      //   `/movements/${this.movement.id}/trees/${this.treeId}/components/parents`,
-      //   {
-      //     [this.subTree]: { subTreeParent: this.parent ? this.parent : "root" },
-      //   }
-      // );
       return setDoc(
         doc(
           getFirestore(),
           `/movements/${this.movement.id}/trees/${this.treeId}/components/parents`
         ),
         {
-          [this.subTree]: { subTreeParent: this.parent ? this.parent : "root" },
+          [this.subTree]: { subTreeParent: this.parent ? this.parent : 'root' },
         },
         { merge: true }
-      ).catch((err) => {
-        // if (process.env.PROD) logEvent(getAnalytics(), 'exception', { description: err, fatal: false })
-        console.error(new Error("Oops, something went wrong: " + err));
-        return false;
-      });
+      )
+        .then(() => {
+          // add the tree to the importedBy tree array
+          setDoc(
+            doc(
+              getFirestore(),
+              `/movements/${this.movement.id}/trees/${this.subTree}`
+            ),
+            {
+              importedBy: arrayUnion(this.treeId),
+            },
+            { merge: true }
+          );
+        })
+        .catch((err) => {
+          // if (process.env.PROD) logEvent(getAnalytics(), 'exception', { description: err, fatal: false })
+          console.error(new Error('Oops, something went wrong: ' + err));
+          return false;
+        });
     },
   },
   computed: {
-    ...mapGetters("movement", ["memberList", "treeOpts"]),
-    ...mapState("movement", ["movement"]),
+    ...mapGetters('movement', ['memberList', 'treeOpts']),
+    ...mapState('movement', ['movement']),
   },
 };
 </script>
