@@ -1,6 +1,6 @@
-exports.users = require("./users.js");
 exports.movements = require("./movements.js");
 exports.members = require("./members.js");
 exports.snapshots = require("./snapshots.js");
 exports.styles = require("./styles.js");
 exports.trees = require("./trees.js");
+exports.users = require("./users.js");
