@@ -1,12 +1,11 @@
 const admin = require("../../../utils/test-admin");
 const environment = require("../../../environments/environment.test.js");
 const context = {
-  admin,
   environment
 };
 
 const Func = require("../acceptInvite.js");
-const db = admin.firestore();
+const db = getFirestore();
 
 const movCol = db.collection(environment.schema.movements);
 
@@ -54,7 +53,7 @@ describe("users: acceptInvite", () => {
     });
 
     it.skip("updates the members in the now doc", async () => {
-      // const date = admin.firestore.Timestamp.fromDate(new Date())
+      // const date = Timestamp.fromDate(new Date())
       // console.log(change)
       await movDoc.set({
         name: "test name",
@@ -153,7 +152,7 @@ describe("users: acceptInvite", () => {
     });
 
     it.skip("updates the stats in the now doc", async () => {
-      // const date = admin.firestore.Timestamp.fromDate(new Date())
+      // const date = Timestamp.fromDate(new Date())
       // console.log(change)
       await movDoc.set({
         name: "test name",
@@ -261,7 +260,7 @@ describe("users: acceptInvite", () => {
 
     // This now happens in a different function
     it.skip("updates the last modified key", async () => {
-      const date = admin.firestore.Timestamp.fromDate(new Date());
+      const date = Timestamp.fromDate(new Date());
       // console.log(change)
       await movDoc
         .collection("styles")
